pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Poorvi Yadav",email = "poorviyadav008@gmail.com"}
]
requires-python = ">=3.11, <4.0"
dependencies = [
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "langchain (>=1.2.12,<2.0.0)",
    "langgraph (>=1.1.2,<2.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "tavily (>=1.1.0,<2.0.0)",
    "chromadb (>=1.5.5,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "langchain-anthropic (>=1.3.5,<2.0.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "langchain-openai (>=1.1.11,<2.0.0)",
    "sentence-transformers (>=5.3.0,<6.0.0)",
    "langchain-huggingface (>=1.2.1,<2.0.0)",
    "langchain-tavily (>=0.2.17,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
